+++ src/Etimo.Id.Api/Controllers/Scopes/ScopeController.cs
@@ -17,12 +17,12 @@ namespace Etimo.Id.Api.Scopes
     [ApiController]
     public class ScopeController : Controller
     {
-        private readonly SiteSettings _siteSettings;
-        private readonly IAddScopeService _addScopeService;
+        private readonly IAddScopeService    _addScopeService;
         private readonly IDeleteScopeService _deleteScopeService;
-        private readonly IFindScopeService _findScopeService;
-        private readonly IGetRolesService _getRolesService;
-        private readonly IGetScopesService _getScopesService;
+        private readonly IFindScopeService   _findScopeService;
+        private readonly IGetRolesService    _getRolesService;
+        private readonly IGetScopesService   _getScopesService;
+        private readonly SiteSettings        _siteSettings;
         private readonly IUpdateScopeService _updateScopeService;
 
         public ScopeController(
@@ -34,12 +34,12 @@ namespace Etimo.Id.Api.Scopes
             IGetScopesService getScopesService,
             IUpdateScopeService updateScopeService)
         {
-            _siteSettings = siteSettings;
-            _addScopeService = addScopeService;
+            _siteSettings       = siteSettings;
+            _addScopeService    = addScopeService;
             _deleteScopeService = deleteScopeService;
-            _findScopeService = findScopeService;
-            _getRolesService = getRolesService;
-            _getScopesService = getScopesService;
+            _findScopeService   = findScopeService;
+            _getRolesService    = getRolesService;
+            _getScopesService   = getScopesService;
             _updateScopeService = updateScopeService;
         }
 
@@ -49,16 +49,10 @@ namespace Etimo.Id.Api.Scopes
         public async Task<IActionResult> GetAsync()
         {
             List<Scope> scopes;
-            if (this.UserHasScope(ScopeScopes.Admin))
-            {
-                scopes = await _getScopesService.GetAllAsync();
-            }
-            else
-            {
-                scopes = await _getScopesService.GetByClientIdAsync(this.GetClientId());
-            }
+            if (this.UserHasScope(ScopeScopes.Admin)) { scopes = await _getScopesService.GetAllAsync(); }
+            else { scopes                                      = await _getScopesService.GetByClientIdAsync(this.GetClientId()); }
 
-            var found = scopes.Select(ScopeResponseDto.FromScope);
+            IEnumerable<ScopeResponseDto> found = scopes.Select(ScopeResponseDto.FromScope);
 
             return Ok(found);
         }
@@ -69,14 +63,8 @@ namespace Etimo.Id.Api.Scopes
         public async Task<IActionResult> FindAsync([FromRoute] Guid scopeId)
         {
             Scope scope;
-            if (this.UserHasScope(ScopeScopes.Admin))
-            {
-                scope = await _findScopeService.FindAsync(scopeId);
-            }
-            else
-            {
-                scope = await _findScopeService.FindAsync(scopeId, this.GetUserId());
-            }
+            if (this.UserHasScope(ScopeScopes.Admin)) { scope = await _findScopeService.FindAsync(scopeId); }
+            else { scope                                      = await _findScopeService.FindAsync(scopeId, this.GetUserId()); }
 
             var found = ScopeResponseDto.FromScope(scope);
 
@@ -89,16 +77,10 @@ namespace Etimo.Id.Api.Scopes
         public async Task<IActionResult> FindRolesAsync([FromRoute] Guid scopeId)
         {
             List<Role> roles;
-            if (this.UserHasScope(ScopeScopes.Admin))
-            {
-                roles = await _getRolesService.GetByScopeIdAsync(scopeId);
-            }
-            else
-            {
-                roles = await _getRolesService.GetByScopeIdAsync(scopeId, this.GetUserId());
-            }
+            if (this.UserHasScope(ScopeScopes.Admin)) { roles = await _getRolesService.GetByScopeIdAsync(scopeId); }
+            else { roles                                      = await _getRolesService.GetByScopeIdAsync(scopeId, this.GetUserId()); }
 
-            var found = roles.Select(r => RoleResponseDto.FromRole(r, false));
+            IEnumerable<RoleResponseDto> found = roles.Select(r => RoleResponseDto.FromRole(r, false));
 
             return Ok(found);
         }
@@ -110,14 +92,8 @@ namespace Etimo.Id.Api.Scopes
         public async Task<IActionResult> CreateAsync([FromBody] ScopeRequestDto dto)
         {
             Scope scope;
-            if (this.UserHasScope(ScopeScopes.Admin))
-            {
-                scope = await _addScopeService.AddAsync(dto.ToScope());
-            }
-            else
-            {
-                scope = await _addScopeService.AddAsync(dto.ToScope(), this.GetUserId());
-            }
+            if (this.UserHasScope(ScopeScopes.Admin)) { scope = await _addScopeService.AddAsync(dto.ToScope()); }
+            else { scope                                      = await _addScopeService.AddAsync(dto.ToScope(), this.GetUserId()); }
 
             var created = ScopeResponseDto.FromScope(scope);
 
@@ -131,14 +107,8 @@ namespace Etimo.Id.Api.Scopes
         public async Task<IActionResult> UpdateAsync([FromRoute] Guid scopeId, [FromBody] ScopeRequestDto dto)
         {
             Scope scope;
-            if (this.UserHasScope(ScopeScopes.Admin))
-            {
-                scope = await _updateScopeService.UpdateAsync(dto.ToScope(scopeId));
-            }
-            else
-            {
-                scope = await _updateScopeService.UpdateAsync(dto.ToScope(scopeId), this.GetUserId());
-            }
+            if (this.UserHasScope(ScopeScopes.Admin)) { scope = await _updateScopeService.UpdateAsync(dto.ToScope(scopeId)); }
+            else { scope = await _updateScopeService.UpdateAsync(dto.ToScope(scopeId), this.GetUserId()); }
 
             var updated = ScopeResponseDto.FromScope(scope);
 
@@ -150,14 +120,8 @@ namespace Etimo.Id.Api.Scopes
         [Authorize(Policy = ScopeScopes.Write)]
         public async Task<IActionResult> DeleteAsync([FromRoute] Guid scopeId)
         {
-            if (this.UserHasScope(ScopeScopes.Admin))
-            {
-                await _deleteScopeService.DeleteAsync(scopeId);
-            }
-            else
-            {
-                await _deleteScopeService.DeleteAsync(scopeId, this.GetUserId());
-            }
+            if (this.UserHasScope(ScopeScopes.Admin)) { await _deleteScopeService.DeleteAsync(scopeId); }
+            else { await _deleteScopeService.DeleteAsync(scopeId, this.GetUserId()); }
 
             return NoContent();
         }
