+++ src/Etimo.Id.Data/Repositories/AccessTokenRepository.cs
@@ -15,26 +15,17 @@ namespace Etimo.Id.Data.Repositories
         }
 
         public Task<AccessToken> FindAsync(Guid accessTokenId)
-        {
-            return _dbContext.AccessTokens.FindAsync(accessTokenId).AsTask();
-        }
+            => _dbContext.AccessTokens.FindAsync(accessTokenId).AsTask();
 
         public void Add(AccessToken token)
-        {
-            _dbContext.AccessTokens.Add(token);
-        }
+            => _dbContext.AccessTokens.Add(token);
 
         public void Delete(AccessToken token)
         {
-            if (token != null)
-            {
-                _dbContext.AccessTokens.Remove(token);
-            }
+            if (token != null) { _dbContext.AccessTokens.Remove(token); }
         }
 
         public Task<int> SaveAsync()
-        {
-            return _dbContext.SaveChangesAsync();
-        }
+            => _dbContext.SaveChangesAsync();
     }
 }
