+++ src/Etimo.Id.Data/Repositories/RoleRepository.cs
@@ -18,46 +18,29 @@ namespace Etimo.Id.Data.Repositories
         }
 
         public Task<List<Role>> GetAllAsync()
-        {
-            return _dbContext.Roles.ToListAsync();
-        }
+            => _dbContext.Roles.ToListAsync();
 
         public Task<List<Role>> GetByUserIdAsync(Guid userId)
-        {
-            return _dbContext.Roles.Where(r => r.Users.Any(u => u.UserId == userId)).ToListAsync();
-        }
+            => _dbContext.Roles.Where(r => r.Users.Any(u => u.UserId == userId)).ToListAsync();
 
         public Task<List<Role>> GetByApplicationId(int applicationId)
-        {
-            return _dbContext.Roles.Where(r => r.ApplicationId == applicationId).ToListAsync();
-        }
+            => _dbContext.Roles.Where(r => r.ApplicationId == applicationId).ToListAsync();
 
         public Task<List<Role>> GetByScopeIdAsync(Guid scopeId)
-        {
-            return _dbContext.Roles.Where(r => r.Scopes.Any(s => s.ScopeId == scopeId)).ToListAsync();
-        }
+            => _dbContext.Roles.Where(r => r.Scopes.Any(s => s.ScopeId == scopeId)).ToListAsync();
 
         public Task<Role> FindAsync(Guid roleId)
-        {
-            return _dbContext.Roles.FindAsync(roleId).AsTask();
-        }
+            => _dbContext.Roles.FindAsync(roleId).AsTask();
 
         public void Add(Role role)
-        {
-            _dbContext.Roles.Add(role);
-        }
+            => _dbContext.Roles.Add(role);
 
         public void Delete(Role role)
         {
-            if (role != null)
-            {
-                _dbContext.Roles.Remove(role);
-            }
+            if (role != null) { _dbContext.Roles.Remove(role); }
         }
 
         public Task<int> SaveAsync()
-        {
-            return _dbContext.SaveChangesAsync();
-        }
+            => _dbContext.SaveChangesAsync();
     }
 }
