+++ src/Etimo.Id.Data/Repositories/ScopeRepository.cs
@@ -3,7 +3,6 @@ using Etimo.Id.Entities;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace Etimo.Id.Data.Repositories
@@ -18,31 +17,20 @@ namespace Etimo.Id.Data.Repositories
         }
 
         public Task<List<Scope>> GetAllAsync()
-        {
-            return _dbContext.Scopes.ToListAsync();
-        }
+            => _dbContext.Scopes.ToListAsync();
 
         public Task<Scope> FindAsync(Guid scopeId)
-        {
-            return _dbContext.Scopes.FindAsync(scopeId).AsTask();
-        }
+            => _dbContext.Scopes.FindAsync(scopeId).AsTask();
 
         public void Add(Scope scope)
-        {
-            _dbContext.Scopes.Add(scope);
-        }
+            => _dbContext.Scopes.Add(scope);
 
         public void Delete(Scope scope)
         {
-            if (scope != null)
-            {
-                _dbContext.Scopes.Remove(scope);
-            }
+            if (scope != null) { _dbContext.Scopes.Remove(scope); }
         }
 
         public Task<int> SaveAsync()
-        {
-            return _dbContext.SaveChangesAsync();
-        }
+            => _dbContext.SaveChangesAsync();
     }
 }
